fix(e2e): rename conversation spec to zzz_ to run after ticket.spec and drop brittle cleanup
